Opdracht1_3Balls_Simple
Created 3 balls, each with different random size, speed and color.
Simplified way of creating balls without the use of a class.

The code used for the random color of the balls is different for each
ball. It demonstrates the difference between setting the color in setup
or in draw.

See comments in code for extra info.